Changes: * change ipxe to use a build directory in the temp-space instead of in the users home directory * change syslinux to use a build directory in the temp-space instead of in the users home directory * change ipxe tasks to:   * always build the monorail.ipxe.debug (see below for reason)   * also build a communications style ipxe.debug. The existing one was more for pci-bus stuff vs looking at dhcp traffic debug.
a note on building the debug images. The contents of static-common has included the hand generated
debug image for awhile. However, since folk didn't know to go and uncomment-out the build for
it, it was not getting updated alongside the actual image. While it is only used in debug
cases, having the two version drift apart so much without any indication caused some
ackward issues in the field.